penambahan beberapa method

+++ src/filemanager/Berkas.java
@@ -13,6 +13,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+import org.apache.commons.io.FileUtils;
 
 /**
  *
@@ -25,6 +26,7 @@ public class Berkas {
   private int jumlahBerkas = 0;
   private long ukuranFile = 0;
   private boolean tersembunyi = false;
+  private ArrayList<Berkas> duplikatanBerkas = new ArrayList<>();
   
   private WebViewUI ui;
   
@@ -38,7 +40,13 @@ public class Berkas {
     // konversikan ukuran file dari bytes ke KB
     this.ukuranFile = objekFile.length() / 1024;
     
-    this.icon = "";
+    if(objekFile.isDirectory()) {
+      this.icon = "assets/Icons/64/101-folder-5.png";
+    }
+    else {
+      this.icon = "assets/Icons/64/053-document-7.png";
+    }
+    
     this.ui = ui;
     
     this.execService = Executors.newCachedThreadPool();
@@ -164,6 +172,99 @@ public class Berkas {
     return hasil.toArray(new Berkas[0]);
   }
   
+  public void hapusKeTrash() throws IOException {
+    ProcessBuilder pb = new ProcessBuilder("gio", "trash", objekFile.getAbsolutePath());
+    Process proc = pb.start();
+    
+    hapusBerkasPadaJS(ui);
+  }
+  
+  public void hapusPermanen() throws IOException {
+    FileUtils.forceDelete(objekFile);
+    
+    hapusBerkasPadaJS(ui);
+  }
+  
+  public Berkas duplikat() throws IOException {
+    Berkas[] listBerkas = new Berkas(ui, objekFile.getParent()).listBerkas();
+    int noSalinan = 1;
+    String namaBerkas = objekFile.getName();
+    
+    for(int i = 0; i < listBerkas.length; i++) {
+      if(listBerkas[i].getObjekFile()
+              .getName().contains(namaBerkas + " (salinan ke-" + noSalinan + ")")) {
+        noSalinan++;
+        i = 0;
+      }
+      else {
+        namaBerkas = objekFile.getName() + " (salinan ke-" + noSalinan + ")";
+      }
+    }
+    
+    Berkas berkasTujuan = new Berkas(ui,
+            objekFile.getParent() + "/" + namaBerkas);
+    Berkas berkasOri = new Berkas(ui, objekFile.getAbsolutePath());
+    
+    if(objekFile.isDirectory()) {
+      FileUtils.copyDirectory(berkasOri.getObjekFile(),
+                              berkasTujuan.getObjekFile());
+    }
+    else {
+      FileUtils.copyFile(berkasOri.getObjekFile(),
+                              berkasTujuan.getObjekFile());
+    }
+    
+    return new Berkas(ui, berkasTujuan.getObjekFile().getAbsolutePath());
+  }
+  
+  public Berkas duplikatDanTampilkan() throws IOException {
+    Berkas berkas = duplikat();
+    berkas.buatBerkasPadaJS();
+    
+    Berkas.tandaiBerkasPadaJS(berkas.getObjekFile().getName(), ui);
+    Berkas.scrollKeBawahPadaJS(ui);
+    
+    return berkas;
+  }
+  
+  public boolean ubahNama(String pathNamaBaru) {
+    boolean sukses = false;
+    
+    if(objekFile.renameTo(new File(pathNamaBaru))) {
+      sukses = true;
+    }
+    
+    return sukses;
+  }
+  
+  public void ubahNamaDanTampilkan(String namaBaru) {
+    String pathNamaBaru = objekFile.getParent() + "/" + namaBaru;
+    
+    if(ubahNama(pathNamaBaru)) {
+      Berkas.ubahNamaPadaJS(ui, namaBaru);
+      Berkas.ubahPathAbsolutPadaJS(ui, pathNamaBaru);
+    }
+  }
+  
+  public void hideBerkas() {
+    String pathNamaBaru = objekFile.getParent() + "/." + objekFile.getName();
+    ubahNama(pathNamaBaru);
+    
+    Berkas.ubahNamaPadaJS(ui, "." + objekFile.getName());
+    Berkas.ubahPathAbsolutPadaJS(ui, pathNamaBaru);
+    Berkas.ubahTersembunyiPadaJS(ui, true);
+  }
+  
+  public void unhideBerkas() {
+    String pathNamaBaru = objekFile.getParent() + "/" +
+                          objekFile.getName().substring(1);
+    ubahNama(pathNamaBaru);
+    
+    Berkas.ubahNamaPadaJS(ui, objekFile.getName().substring(1));
+    Berkas.ubahPathAbsolutPadaJS(ui, pathNamaBaru);
+    Berkas.ubahTersembunyiPadaJS(ui, false);
+  }
+  
   public static Berkas buatFolderBaru(String pathFolder, WebViewUI ui)
           throws AccessDeniedException, IOException, FileAlreadyExistsException {
     Berkas berkas = null;
@@ -193,26 +294,79 @@ public class Berkas {
     return berkas;
   }
   
-  public static boolean pindahkanBerkas(Berkas berkas, Berkas tujuan)
+  public static boolean pindahkanFile(WebViewUI ui, Berkas berkas, Berkas tujuan)
           throws IOException {
     
     boolean sukses = false;
     
-    Path op = Files.move(berkas.getObjekFile().toPath(),
-            berkas.getObjekFile().toPath());
-    
-    if(op != null) {
-      sukses = true;
+    if(berkas.getObjekFile().isFile()) {
+      buatPanelOpPadaJS(ui, "", "pemindahan",
+              berkas.getObjekFile().getAbsolutePath(),
+              tujuan.getObjekFile().getAbsolutePath());
+      
+      Path op = Files.move(berkas.getObjekFile().toPath(),
+              berkas.getObjekFile().toPath());
+
+      if(op != null) {
+        sukses = true;
+      }
     }
     
     return sukses;
+  }
+  
+  public static void buatPanelOpPadaJS(WebViewUI ui,
+          String var, String op, String pathBerkas, String pathTujuan) {
+    
+    String js = ""+
+    "var "+var+" = new PanelOperasiBerkas('"+op+"')"+
+    ".setPathAktif('"+pathBerkas+"')"+
+    ".setPathTujuan('"+pathTujuan+"');"+
+    ".pasangElemen($('#rowOperasiBerkas'));";
+    
+    ui.eksekusiJavascript(js);
+  }
+  
+  public static void ubahPathAktif(WebViewUI ui, String var, String path) {
+    String js = ""+
+    var+".ubahPathAktif('"+path+"');";
+    
+    ui.eksekusiJavascript(js);
+  }
+  
+  public void hapusBerkasPadaJS(WebViewUI ui) {
+    String js = ""+
+    "Berkas.hapusBerkasBerdasarNama('"+objekFile.getName()+"')";
+    
+    ui.eksekusiJavascript(js);
+  }
+  
+  public static void ubahTersembunyiPadaJS(WebViewUI ui, boolean sembunyikan) {
+    String js = ""+
+    "Berkas.dapatkanBerkasTerpilih().ubahTersembunyi("+sembunyikan+");";
+    
+    ui.eksekusiJavascript(js);
+  }
+  
+  public static void ubahPathAbsolutPadaJS(WebViewUI ui, String path) {
+    String js = ""+
+    "Berkas.dapatkanBerkasTerpilih().ubahPathAbsolut('"+path+"');";
+    
+    ui.eksekusiJavascript(js);
+  }
+  
+  public static void ubahNamaPadaJS(WebViewUI ui, String nama) {
+    String js = ""+
+    "Berkas.dapatkanBerkasTerpilih().ubahNama('"+nama+"');";
+    
+    ui.eksekusiJavascript(js);
   }
   
   public void hapusSemuaBerkasPadaJS() {
     ui.eksekusiJavascript("Berkas.hapusSemuaBerkas();");
   }
   
-  public void tampilkanCirclePadaJS() {
+  public static void tampilkanCirclePadaJS(WebViewUI ui) {
     String js = ""+
     "$('#konten').hide();"+
     "$('#loadingCircle').show();";
@@ -220,7 +374,7 @@ public class Berkas {
     ui.eksekusiJavascript(js);
   }
   
-  public void sembunyikanCirclePadaJS() {
+  public static void sembunyikanCirclePadaJS(WebViewUI ui) {
     String js = ""+
     "$('#konten').show();"+
     "$('#loadingCircle').hide();";
@@ -232,6 +386,20 @@ public class Berkas {
     String js = "" +
     "Berkas.hilangkanSemuaTanda();"+
     "Berkas.tandai('"+namaBerkas+"');";
+    
+    ui.eksekusiJavascript(js);
+  }
+  
+  public static void setTextPathPadaJS(WebViewUI ui, String path) {
+    String js = ""+
+    "$('#txPath').val('"+path+"');";
+    
+    ui.eksekusiJavascript(js);
+  }
+  
+  public static void scrollKeBawahPadaJS(WebViewUI ui) {
+    String js = ""+
+    "$('#konten').animate({ scrollTop: $(document).height() }, 500);";
     
     ui.eksekusiJavascript(js);
   }
